Testability/OrderProcessor.cs
using System;

namespace Testability
{
    public class OrderProcessor
    {
        // Example of loose coupling

        private readonly IShippingCalculator _shippingCalculator;

        public OrderProcessor(IShippingCalculator shippingCalculator)
        {
            _shippingCalculator = shippingCalculator;
        }

        public void Process(Order order)
        {
            // Defensive Programming: Add some checks, but strike a balance
            // between checks and bloated code

            if (order.IsShipped)
                throw new InvalidOperationException("This order is already processed.");

            order.Shipment = new Shipment
            {
                ShippingCost = _shippingCalculator.CalculateShipping(order),
                ShippingDate = DateTime.Today.AddDays(1)
            };
        }
    }
}